% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Scales a numeric vector to a [0, 1] range.}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Returns the scaled vector, all values in the 0-1 range. If all the values in the vector are same it will return 0.5 which is the mid value
}
\description{
This function takes a vector and performs min-max normalization
}
\examples{
# example 1
vec = sample(1:100,20)
min_max_scale(vec)

# example 2
vec = rep(5, 10)
min_max_scale(vec)

}
